Add uruguay records and update uruguay releases
